webspeech api support

+++ src/containers/WebSpeechContainer.js
@@ -1,24 +1,36 @@
 const util = require('util')
-const fs = require('fs')
 const opn = require('opn')
 const path = require('path')
 const async = require('async')
-const http = require('http')
+const findRoot = require('find-root')
+const HttpServer = require('tiny-server')
 const socket = require('socket.io')
 const debug = require('debug')('botium-WebSpeechContainer')
 
+const Capabilities = require('../Capabilities')
 const Events = require('../Events')
 const BaseContainer = require('./BaseContainer')
 const BotiumMockMessage = require('../mocks/BotiumMockMessage')
 
+const botiumPackageRootDir = findRoot()
+
 module.exports = class WebSpeechContainer extends BaseContainer {
   Validate () {
     return super.Validate().then(() => {
-      // this._AssertCapabilityExists(Capabilities.BOTIUMGRIDURL)
+      this._AssertCapabilityExists(Capabilities.WEBSPEECH_SERVER_PORT)
+      this._AssertCapabilityExists(Capabilities.WEBSPEECH_LANGUAGE)
     })
   }
 
   Build () {
+    this.browserConfig = {
+      WEBSPEECH_LANGUAGE: this.caps[Capabilities.WEBSPEECH_LANGUAGE],
+      WEBSPEECH_PITCH: this.caps[Capabilities.WEBSPEECH_PITCH],
+      WEBSPEECH_RATE: this.caps[Capabilities.WEBSPEECH_RATE],
+      WEBSPEECH_VOLUME: this.caps[Capabilities.WEBSPEECH_VOLUME],
+      WEBSPEECH_VOICE: this.caps[Capabilities.WEBSPEECH_VOICE]
+    }
+
     return new Promise((resolve, reject) => {
       async.series([
         (baseComplete) => {
@@ -26,13 +38,7 @@ module.exports = class WebSpeechContainer extends BaseContainer {
         },
 
         (httpServerComplete) => {
-          this.httpServer = http.createServer((req, res) => {
-            debug('serving html file')
-            res.writeHead(200, {
-              'Context-Type': 'text/html'
-            })
-            fs.createReadStream(path.resolve(__dirname, 'WebSpeechContainer.html')).pipe(res)
-          })
+          this.httpServer = new HttpServer(path.resolve(botiumPackageRootDir, 'src', 'containers', 'webspeech'))
           this.io = socket(this.httpServer)
           this.io.on('connection', (clientSocket) => {
             debug('browser connected to socket')
@@ -44,10 +50,16 @@ module.exports = class WebSpeechContainer extends BaseContainer {
             this.clientSocket.on('log', (msg) => {
               debug(`browser log: ${msg}`)
             })
-            this.clientSocket.on('listened', (msg) => {
-              debug(`browser listened: ${msg}`)
+            this.clientSocket.on('botsays', (msg) => {
+              debug(`browser botsays: ${msg}`)
               if (msg) {
                 this._QueueBotSays(new BotiumMockMessage({ messageText: msg }))
+              }
+            })
+            this.clientSocket.on('usersaid', (msg) => {
+              if (this.usersaidResolve) {
+                this.usersaidResolve()
+                this.usersaidResolve = null
               }
             })
 
@@ -56,15 +68,20 @@ module.exports = class WebSpeechContainer extends BaseContainer {
               this.connectResolve = null
             }
           })
-          this.httpServer.listen(3000, () => {
-            debug('waiting for browser connection')
+          this.httpServer.listen(this.caps[Capabilities.WEBSPEECH_SERVER_PORT], () => {
+            debug(`waiting for browser connection on port ${this.caps[Capabilities.WEBSPEECH_SERVER_PORT]}`)
             httpServerComplete()
+          }).once('error', (err) => {
+            if (err.code === 'EADDRINUSE') {
+              debug(`port ${this.caps[Capabilities.WEBSPEECH_SERVER_PORT]} already in use.`)
+              httpServerComplete(err)
+            }
           })
         }
 
       ], (err) => {
         if (err) {
-          return reject(new Error(`Cannot build simplereset container: ${util.inspect(err)}`))
+          return reject(new Error(`Cannot build webspeech container: ${util.inspect(err)}`))
         }
         resolve(this)
       })
@@ -85,7 +102,15 @@ module.exports = class WebSpeechContainer extends BaseContainer {
             opnComplete()
           } else {
             this.connectResolve = opnComplete
-            opn('http://127.0.0.1:3000', { app: 'Chrome' }).then((cp) => {
+
+            let opnOptions = { }
+            if (this.caps[Capabilities.WEBSPEECH_BROWSER_APP]) {
+              opnOptions.app = this.caps[Capabilities.WEBSPEECH_BROWSER_APP]
+            }
+
+            const browserUrl = `http://127.0.0.1:${this.caps[Capabilities.WEBSPEECH_SERVER_PORT]}/WebSpeechContainer.html`
+            debug(`opening browser process to point to url ${browserUrl}`)
+            opn(browserUrl, opnOptions).then((cp) => {
               debug(`browser process running`)
               this.clientProcess = cp
             })
@@ -106,9 +131,11 @@ module.exports = class WebSpeechContainer extends BaseContainer {
   UserSays (mockMsg) {
     return new Promise((resolve, reject) => {
       if (this.clientSocket) {
-        this.clientSocket.emit('speak', mockMsg)
-        this.eventEmitter.emit(Events.MESSAGE_SENTTOBOT, this, mockMsg)
-        resolve(this)
+        this.usersaidResolve = () => {
+          this.eventEmitter.emit(Events.MESSAGE_SENTTOBOT, this, mockMsg)
+          resolve(this)
+        }
+        this.clientSocket.emit('usersays', this.browserConfig, mockMsg)
       } else {
         this.eventEmitter.emit(Events.MESSAGE_SENDTOBOT_ERROR, this, 'browser connection not online')
         reject(new Error('browser connection not online'))
@@ -118,7 +145,7 @@ module.exports = class WebSpeechContainer extends BaseContainer {
 
   WaitBotSays (channel = null, timeoutMillis = null) {
     if (this.clientSocket) {
-      this.clientSocket.emit('listen')
+      this.clientSocket.emit('waitbotsays', this.browserConfig)
       return super.WaitBotSays(channel, timeoutMillis)
     } else {
       return Promise.reject(new Error('browser connection not online'))
@@ -157,10 +184,12 @@ module.exports = class WebSpeechContainer extends BaseContainer {
         (httpServerStopped) => {
           this.io = null
           if (this.clientSocket) {
-            this.clientSocket.emit('close')
+            if (this.caps[Capabilities.WEBSPEECH_CLOSEBROWSER]) {
+              this.clientSocket.emit('close')
+            }
             this.clientSocket.disconnect(true)
+            this.clientSocket = null
           }
-          this.clientSocket = null
           if (this.httpServer) {
             debug('closing http server')
             this.httpServer.close(() => {
@@ -169,17 +198,18 @@ module.exports = class WebSpeechContainer extends BaseContainer {
               httpServerStopped()
             })
           } else {
-            this.httpServer = null
             httpServerStopped()
           }
         },
 
         (clientProcessStopped) => {
-          if (this.clientProcess) {
-            debug('killing browser process')
-            this.clientProcess.kill()
-            this.clientProcess = null
+          if (this.caps[Capabilities.WEBSPEECH_CLOSEBROWSER]) {
+            if (this.clientProcess) {
+              debug('killing browser process')
+              this.clientProcess.kill()
+            }
           }
+          this.clientProcess = null
           clientProcessStopped()
         }
